Move ethereum browser wallet sendTransaction logic to ethers.ts
